spatial-audio/pyproject.toml
[project]
name = "spatial-audio"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Noam Shabtai", email = "shabtai.noam@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "activator",
    "audio-io",
    "coordinates",
    "numpy>=2.3.5",
    "numpy-quaternion>=2024.0.12",
    "stft",
]

[project.scripts]
spatial-audio = "spatial_audio:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "parametrize-tests",
    "pytest>=9.0.1",
]

[tool.uv.sources]
parametrize-tests = { path = "../parametrize-tests", editable = true }
audio-io = { path = "../audio-io", editable = true }
coordinates = { path = "../coordinates", editable = true }
stft = { path = "../stft", editable = true }
activator = { path = "../activator", editable = true }
